requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.1
djangorestframework==3.8.2
Faker==0.8.17
future==0.16.0
idna==2.7
oauthlib==2.1.0
Pillow==5.2.0
PyJWT==1.6.4
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
text-unidecode==1.2
urllib3==1.23
uWSGI==2.0.17.1
